# PERF: Update Autoscoper to include PSO optimization